v5.2 WiFi reconnect & rewrite values save
The OpenCO2 Sensor now auto reconnects if the WiFi connection is lost (for example a router reboot).

Code refactoring: Rewrite how values are stored. Now up to 24h of measurements are stored in a ring buffer with automatic overflow handling. This replaces the more complex handling of 24 separate arrays containing measurements for 1h each.