v6.0 increase batterylife
New default to low energy mode with a smart algorithm to only measure CO2, if the temperature changed by more than 0.5°C or the humidity by 2% or 5 minutes passed. Increase battery runtime to up to 11 weeks.

Bugfix: LED color fixed on HWSubRev < 3

Also USB updates are only enabeled if the OpenCO2 Sensor is plugged in already on the first boot.